Extract toolbar format button components
