+++ libs/designer-ui/src/lib/html/plugins/toolbar/Format.tsx
@@ -1,31 +1,24 @@
 import constants from '../../../constants';
 import fontColorSvgDark from '../icons/dark/font-color.svg';
-import linkDark from '../icons/dark/link.svg';
 import paintBucketSvgDark from '../icons/dark/paint-bucket.svg';
-import boldDark from '../icons/dark/type-bold.svg';
-import italicDark from '../icons/dark/type-italic.svg';
-import underlineDark from '../icons/dark/type-underline.svg';
 import fontColorSvgLight from '../icons/light/font-color.svg';
-import linkLight from '../icons/light/link.svg';
 import paintBucketSvgLight from '../icons/light/paint-bucket.svg';
-import boldLight from '../icons/light/type-bold.svg';
-import italicLight from '../icons/light/type-italic.svg';
-import underlineLight from '../icons/light/type-underline.svg';
+import { FormatBoldButton } from './buttons/FormatBoldButton';
+import { FormatItalicButton } from './buttons/FormatItalicButton';
+import { FormatLinkButton } from './buttons/FormatLinkButton';
+import { FormatUnderlineButton } from './buttons/FormatUnderlineButton';
 import { DropdownColorPicker } from './DropdownColorPicker';
 import { getSelectedNode, sanitizeUrl } from './helper/functions';
 import { useTheme } from '@fluentui/react';
-import { ToolbarButton } from '@fluentui/react-components';
 import { $isLinkNode, TOGGLE_LINK_COMMAND } from '@lexical/link';
 import { $getSelectionStyleValueForProperty, $patchStyleText } from '@lexical/selection';
 import { mergeRegister } from '@lexical/utils';
-import { isApple } from '@microsoft/logic-apps-shared';
 import type { LexicalEditor } from 'lexical';
 import {
   $getSelection,
   $isRangeSelection,
   COMMAND_PRIORITY_CRITICAL,
   COMMAND_PRIORITY_NORMAL,
-  FORMAT_TEXT_COMMAND,
   KEY_MODIFIER_COMMAND,
   SELECTION_CHANGE_COMMAND,
 } from 'lexical';
@@ -140,81 +133,6 @@ export const Format = ({ activeEditor, readonly }: FormatProps) => {
     [applyStyleText]
   );
 
-  const insertLink = useCallback(() => {
-    if (isLink) {
-      activeEditor.dispatchCommand(TOGGLE_LINK_COMMAND, null);
-    } else {
-      activeEditor.dispatchCommand(TOGGLE_LINK_COMMAND, sanitizeUrl('https://'));
-    }
-  }, [activeEditor, isLink]);
-
-  const boldTitleMac = intl.formatMessage({
-    defaultMessage: 'Bold (⌘B)',
-    id: 'ciLkfU',
-    description: 'Command for bold text for Mac users',
-  });
-  const boldTitleMacAriaLabel = intl.formatMessage({
-    defaultMessage: 'Format text as bold. Shortcut: ⌘B',
-    id: 'S138/4',
-    description: 'label to make bold text for Mac users',
-  });
-  const boldTitleNonMac = intl.formatMessage({
-    defaultMessage: 'Bold (Ctrl+B)',
-    id: 'Lnqh6h',
-    description: 'Command for bold text for non-mac users',
-  });
-  const boldTitleNonMacAriaLabel = intl.formatMessage({
-    defaultMessage: 'Format text as bold. Shortcut: Ctrl+B',
-    id: 'YR1uWE',
-    description: 'label to make bold text for nonMac users',
-  });
-
-  const italicTitleMac = intl.formatMessage({
-    defaultMessage: 'Italic (⌘I)',
-    id: 'elD6+N',
-    description: 'Command for italic text for Mac users',
-  });
-  const italicTitleMacAriaLabel = intl.formatMessage({
-    defaultMessage: 'Format text as italic. Shortcut: ⌘I',
-    id: 'YdQw4/',
-    description: 'label to make italic text for Mac users',
-  });
-  const italicTitleNonMac = intl.formatMessage({
-    defaultMessage: 'Italic (Ctrl+I)',
-    id: 'dfmH55',
-    description: 'Command for italic text for non-mac users',
-  });
-  const italicTitleNonMacAriaLabel = intl.formatMessage({
-    defaultMessage: 'Format text as italic. Shortcut: Ctrl+I',
-    id: 'gIx5ys',
-    description: 'label to make italic text for nonMac users',
-  });
-
-  const underlineTitleMac = intl.formatMessage({
-    defaultMessage: 'Underline (⌘U)',
-    id: 'KYX5Do',
-    description: 'Command for underline text for Mac users',
-  });
-  const underlineTitleMacAriaLabel = intl.formatMessage({
-    defaultMessage: 'Format text as underline. Shortcut: ⌘U',
-    id: 'qBkxGU',
-    description: 'label to make underline text for Mac users',
-  });
-  const underlineTitleNonMac = intl.formatMessage({
-    defaultMessage: 'Underline (Ctrl+U)',
-    id: 'lwlg2K',
-    description: 'Command for underline text for non-mac users',
-  });
-  const underlineTitleNonMacAriaLabel = intl.formatMessage({
-    defaultMessage: 'Format text as underline. Shortcut: Ctrl+U',
-    id: 'YJlS8E',
-    description: 'label to make underline text for nonMac users',
-  });
-  const insertLinkLabel = intl.formatMessage({
-    defaultMessage: 'Insert Link',
-    id: 'tUCptx',
-    description: 'label to insert link',
-  });
   const backgroundColorTitle = intl.formatMessage({
     defaultMessage: 'Background Color',
     id: 'r7ZizR',
@@ -228,39 +146,9 @@ export const Format = ({ activeEditor, readonly }: FormatProps) => {
 
   return (
     <>
-      <ToolbarButton
-        onMouseDown={(e) => e.preventDefault()}
-        onClick={() => {
-          activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, 'bold');
-        }}
-        className={`toolbar-item spaced ${isBold ? 'active' : ''}`}
-        title={isApple() ? boldTitleMac : boldTitleNonMac}
-        aria-label={isApple() ? boldTitleMacAriaLabel : boldTitleNonMacAriaLabel}
-        disabled={readonly}
-        icon={<img className={'format'} src={isInverted ? boldDark : boldLight} alt={'bold icon'} />}
-      />
-      <ToolbarButton
-        onMouseDown={(e) => e.preventDefault()}
-        onClick={() => {
-          activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, 'italic');
-        }}
-        className={`toolbar-item spaced ${isItalic ? 'active' : ''}`}
-        title={isApple() ? italicTitleMac : italicTitleNonMac}
-        aria-label={isApple() ? italicTitleMacAriaLabel : italicTitleNonMacAriaLabel}
-        disabled={readonly}
-        icon={<img className={'format'} src={isInverted ? italicDark : italicLight} alt={'italic icon'} />}
-      />
-      <ToolbarButton
-        onMouseDown={(e) => e.preventDefault()}
-        onClick={() => {
-          activeEditor.dispatchCommand(FORMAT_TEXT_COMMAND, 'underline');
-        }}
-        className={`toolbar-item spaced ${isUnderline ? 'active' : ''}`}
-        title={isApple() ? underlineTitleMac : underlineTitleNonMac}
-        aria-label={isApple() ? underlineTitleMacAriaLabel : underlineTitleNonMacAriaLabel}
-        disabled={readonly}
-        icon={<img className={'format'} src={isInverted ? underlineDark : underlineLight} alt={'underline icon'} />}
-      />
+      <FormatBoldButton activeEditor={activeEditor} isToggledOn={isBold} readonly={readonly} />
+      <FormatItalicButton activeEditor={activeEditor} isToggledOn={isItalic} readonly={readonly} />
+      <FormatUnderlineButton activeEditor={activeEditor} isToggledOn={isUnderline} readonly={readonly} />
       <DropdownColorPicker
         editor={activeEditor}
         disabled={readonly}
@@ -281,15 +169,7 @@ export const Format = ({ activeEditor, readonly }: FormatProps) => {
         onChange={onBgColorSelect}
         title={backgroundColorTitle}
       />
-      <ToolbarButton
-        onMouseDown={(e) => e.preventDefault()}
-        disabled={readonly}
-        onClick={insertLink}
-        className={`toolbar-item spaced ${isLink ? 'active' : ''}`}
-        aria-label={insertLinkLabel}
-        title={insertLinkLabel}
-        icon={<img className={'format'} src={isInverted ? linkDark : linkLight} alt={'link icon'} />}
-      />
+      <FormatLinkButton activeEditor={activeEditor} isToggledOn={isLink} readonly={readonly} />
     </>
   );
 };
